Update
1. Move is_hdd() declaration from util.h into is_hdd.h
  include "common/is_hdd.h" into *.cpp that use is_hdd()
2. Move the following functions from util.cpp into is_hdd.cpp
  is_hdd_sysfs()
  fp2lv()
  lv2pv()
  win_device_has_seek_penalty()
  is_hdd_win_ioctl()
  is_hdd()
3. Fix the following things
  use _WIN32 instead of WIN32
  use unsigned short instead of uint16_t
  use penalty instead of penalny
  add CloseHandle(h) in all conditional branches to avoid resource leaking
  use either 0/1 or true/false to maintain consistency
  rename tests/unit_tests/test_a.cpp
