+++ src/common/util.h
@@ -230,10 +230,6 @@ namespace tools
   bool sha256sum(const uint8_t *data, size_t len, crypto::hash &hash);
   bool sha256sum(const std::string &filename, crypto::hash &hash);
 
-
-  bool is_hdd(const char *path, bool &result);
-  bool is_hdd(const char *path);
-
   boost::optional<std::pair<uint32_t, uint32_t>> parse_subaddress_lookahead(const std::string& str);
 
   std::string glob_to_regex(const std::string &val);
