+++ tests/unit_tests/CMakeLists.txt
@@ -72,7 +72,7 @@ set(unit_tests_sources
   ringct.cpp
   output_selection.cpp
   vercmp.cpp
-  test_a.cpp
+  is_hdd.cpp
   ringdb.cpp)
 
 set(unit_tests_headers
